HT-16/basic/requirements.txt
chromedriver-autoinstaller==0.6.3
selenium==4.16.0
Pillow==10.1.0
requests==2.31.0
